Refactor extract helper
